requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12

# Database
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1
pydantic[email]==2.9.2

# Machine Learning
tensorflow==2.18.0
numpy==1.26.4
scikit-learn==1.5.2
joblib==1.4.2
pandas==2.2.3
opencv-python==4.10.0.84
pillow==10.4.0

# Optional: For better performance
# gunicorn==23.0.0
